backend/requirements-prod.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.1
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
django-filter==23.5
Pillow==10.2.0
gunicorn==21.2.0
whitenoise==6.6.0